removed references to long() as Python 3.x does not use these anymore

+++ hx711.py
@@ -30,7 +30,7 @@ class HX711:
 
         self.OFFSET = 1
         self.OFFSET_B = 1
-        self.lastVal = long(0)
+        self.lastVal = 0
 
         self.DEBUG_PRINTING = False
 
@@ -161,7 +161,7 @@ class HX711:
         self.lastVal = signedIntValue
 
         # Return the sample value we've read from the HX711.
-        return long(signedIntValue)
+        return signedIntValue
 
     
     def read_average(self, times=3):
